Add support for mongo as the backend DB
docker-compose files specifies a mongodb container.
This file is generated by my version of `prisma init` which
is slightly outdated (1.32.0-beta).  I was originally using
that version because the hosted prisma demo in the us-west
zone was running that version. But I could probably now
run the latest release of prisma CLI.